adding equals overloads that allow you to specify the epsilon (for unit tests); switching to double precision for weights everywhere

+++ src/main/java/com/github/steveash/jopenfst/Arc.java
@@ -21,7 +21,7 @@ package com.github.steveash.jopenfst;
  */
 public interface Arc {
 
-  float getWeight();
+  double getWeight();
 
   int getIlabel();
