+++ src/main/java/com/github/steveash/jopenfst/semiring/LogSemiring.java
@@ -28,10 +28,10 @@ public class LogSemiring extends Semiring {
   private static final long serialVersionUID = 5212106775584311083L;
 
   // zero value
-  private static float zero = Float.POSITIVE_INFINITY;
+  private static double zero = Double.POSITIVE_INFINITY;
 
   // one value
-  private static float one = 0.f;
+  private static double one = 0.0;
 
   /*
    * (non-Javadoc)
@@ -41,16 +41,16 @@ public class LogSemiring extends Semiring {
    * edu.cmu.sphinx.fst.weight.float)
    */
   @Override
-  public float plus(float w1, float w2) {
+  public double plus(double w1, double w2) {
     if (!isMember(w1) || !isMember(w2)) {
-      return Float.NEGATIVE_INFINITY;
+      return Double.NEGATIVE_INFINITY;
     }
-    if (w1 == Float.POSITIVE_INFINITY) {
+    if (w1 == Double.POSITIVE_INFINITY) {
       return w2;
-    } else if (w2 == Float.POSITIVE_INFINITY) {
+    } else if (w2 == Double.POSITIVE_INFINITY) {
       return w1;
     }
-    return (float) -Math.log(Math.exp(-w1) + Math.exp(-w2));
+    return (double) -Math.log(Math.exp(-w1) + Math.exp(-w2));
   }
 
   /*
@@ -61,9 +61,9 @@ public class LogSemiring extends Semiring {
    * edu.cmu.sphinx.fst.weight.float)
    */
   @Override
-  public float times(float w1, float w2) {
+  public double times(double w1, double w2) {
     if (!isMember(w1) || !isMember(w2)) {
-      return Float.NEGATIVE_INFINITY;
+      return Double.NEGATIVE_INFINITY;
     }
 
     return w1 + w2;
@@ -77,13 +77,13 @@ public class LogSemiring extends Semiring {
    * , edu.cmu.sphinx.fst.weight.float)
    */
   @Override
-  public float divide(float w1, float w2) {
+  public double divide(double w1, double w2) {
     if (!isMember(w1) || !isMember(w2)) {
-      return Float.NEGATIVE_INFINITY;
+      return Double.NEGATIVE_INFINITY;
     }
 
     if (w2 == zero) {
-      return Float.NEGATIVE_INFINITY;
+      return Double.NEGATIVE_INFINITY;
     } else if (w1 == zero) {
       return zero;
     }
@@ -97,7 +97,7 @@ public class LogSemiring extends Semiring {
    * @see edu.cmu.sphinx.fst.weight.Semiring#zero()
    */
   @Override
-  public float zero() {
+  public double zero() {
     return zero;
   }
 
@@ -107,7 +107,7 @@ public class LogSemiring extends Semiring {
    * @see edu.cmu.sphinx.fst.weight.Semiring#one()
    */
   @Override
-  public float one() {
+  public double one() {
     return one;
   }
 
@@ -119,9 +119,9 @@ public class LogSemiring extends Semiring {
    * .float)
    */
   @Override
-  public boolean isMember(float w) {
-    return (!Float.isNaN(w)) // not a NaN
-           && (w != Float.NEGATIVE_INFINITY); // and different from -inf
+  public boolean isMember(double w) {
+    return (!Double.isNaN(w)) // not a NaN
+           && (w != Double.NEGATIVE_INFINITY); // and different from -inf
   }
 
   /*
@@ -130,7 +130,7 @@ public class LogSemiring extends Semiring {
    * @see edu.cmu.sphinx.fst.semiring.Semiring#reverse(float)
    */
   @Override
-  public float reverse(float w1) {
+  public double reverse(double w1) {
     throw new UnsupportedOperationException("not implemented");
 //    return Float.NEGATIVE_INFINITY;
   }
